Add files via upload
All the files that have been uploaded are files involved in the data based project. These were the files given by the competition. The PDF shows the final insights from the data and the code is given in Q1 and Q2. This was a team based project and as such, this is not all the code that was made.